config-backup/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "config-backup"
version = "0.1.0"
description = "A CLI tool to automatically backup configuration files on Ubuntu Server"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
config-backup = "config_backup.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["config_backup*"]
